misc: simplify code and comment
